✨ feat(ArtistCard.tsx): 아티스트 카드 전역으로 쓸 수 있도록 input 기능과 분리
